requirements.txt
langchain
langchain-ollama
langchain_community
streamlit
tiktoken
beautifulsoup4
markitdown[all]
docling
faiss-cpu